💻️ feat: Add copyToClipboard utility function and use it in OptionWebhookUrl component
