+++ assets/react/v3/shared/utils/util.ts
@@ -329,3 +329,29 @@ export const convertGMTtoLocalDate = (date: string) => {
 export const normalizeLineEndings = (text: string) => {
   return (text || '').replace(/\r\n/g, '\n');
 };
+
+export const copyToClipboard = (text: string) => {
+  return new Promise<void>((resolve, reject) => {
+    if (navigator.clipboard && window.isSecureContext) {
+      navigator.clipboard
+        .writeText(text)
+        .then(() => resolve())
+        .catch((error) => reject(error));
+    } else {
+      const textarea = document.createElement('textarea');
+      textarea.value = text;
+      document.body.appendChild(textarea);
+      textarea.select();
+
+      try {
+        // if navigator.clipboard is not available, use document.execCommand('copy')
+        document.execCommand('copy');
+        resolve();
+      } catch (error) {
+        reject(error);
+      } finally {
+        document.body.removeChild(textarea); // Clean up
+      }
+    }
+  });
+};
